requirements.txt
astroid==2.3.3
attrs==19.3.0
Click==7.0
dependency-injector==3.15.6
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
importlib-metadata==1.5.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpath==0.82
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dotenv==0.12.0
rfc3339-validator==0.1.2
Shapely==1.7.0
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
xmlunittest==0.5.0
zipp==3.1.0
